:recycle: refactor(recipe): migrate to ESM next.config + shared CSP (#559)
Convert next.config.js to .mjs and consume @howardism/security-headers
with the default CSP directives — https: in img-src and connect-src
already covers Cloudinary and the CMS axios calls.
