--- apps/recipe/next.config.js
@@ -1,36 +1,0 @@
-const withBundleAnalyzer = require("@next/bundle-analyzer")({
-  enabled: process.env.ANALYZE === "true",
-});
-
-const securityHeaders = [
-  { key: "X-Content-Type-Options", value: "nosniff" },
-  { key: "X-Frame-Options", value: "DENY" },
-  { key: "X-XSS-Protection", value: "0" },
-  { key: "Referrer-Policy", value: "strict-origin-when-cross-origin" },
-  {
-    key: "Strict-Transport-Security",
-    value: "max-age=63072000; includeSubDomains",
-  },
-  {
-    key: "Permissions-Policy",
-    value: "camera=(), microphone=(), geolocation=()",
-  },
-];
-
-/** @type{import('next').NextConfig} */
-const config = {
-  reactStrictMode: true,
-  async headers() {
-    return [{ source: "/(.*)", headers: securityHeaders }];
-  },
-  images: {
-    remotePatterns: [{ hostname: "res.cloudinary.com" }],
-  },
-  transpilePackages: [
-    "@howardism/components-common",
-    "@howardism/login-form",
-    "@howardism/ui",
-  ],
-};
-
-module.exports = withBundleAnalyzer(config);
