+++ apps/recipe/package.json
@@ -20,6 +20,7 @@
     "@hookform/resolvers": "^5.2.2",
     "@howardism/components-common": "workspace:*",
     "@howardism/login-form": "workspace:*",
+    "@howardism/security-headers": "workspace:*",
     "@howardism/ui": "workspace:*",
     "@reduxjs/toolkit": "^2.2.1",
     "axios": "^1.6.7",
